Update README.md
Yea so I kinda went a bit over the top and kinda re-wrote a lot of it. Take what you like, disregard what you like or disregard it all if you wish. There are a few formatting changes where the hugo instructions have been added to better align with your original additions. I am not sure 100 on the method as I have not made one but it will be a good idea to still get a set of eyes on this if you decide to use parts.